+++ src/integration_tests/incremental_adding_of_students.py
@@ -1,0 +1,3 @@
+from eksam.testing import server_running
+
+
